ragtube/requirements_py313.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pinecone-client>=3.0.0
pinecone-text>=0.7.0
sentence-transformers>=2.2.0
cohere>=4.30.0
youtube-transcript-api>=0.6.0
google-api-python-client>=2.100.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.0
pydantic>=2.5.0
torch>=2.1.0
transformers>=4.30.0
tiktoken>=0.5.0
PyYAML>=6.0.0
